% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{add_multicol}
\alias{add_multicol}
\title{produce a factor with 'positive' cell type values for SpatialExperiment annotated like EH7312}
\usage{
add_multicol(spe, inds = 188:194, opts = c("CD68", "CK", "CD19", "CD3", "CD8"))
}
\arguments{
\item{spe}{SpatialExperiment instance}

\item{inds}{numeric() indices of 'phenotype' variables in `colData(spe)`}

\item{opts}{character() prefixes of values for phenotype values}
}
\description{
produce a factor with 'positive' cell type values for SpatialExperiment annotated like EH7312
}
\note{
This is a speculative implementation.  If multiple factors for a cell
are positive, the last one (in the ordering of `opts`) is used.
}
\examples{
requireNamespace("SummarizedExperiment")
data(litov)
if ("multicol" \%in\% names(SummarizedExperiment::colData(litov))) {
  wh = which(names(SummarizedExperiment::colData(litov)) == "multicol")
  colData(litov) = SummarizedExperiment::colData(litov)[,-wh]
  }
litov = add_multicol(litov)
table(litov$multicol, exclude=NULL)
}
