% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{display_by_core}
\alias{display_by_core}
\title{Explore ovVP Vectra Polaris data from ExperimentHub EH7312}
\usage{
display_by_core(spe, phv, add_cd4 = TRUE, add_multicol = TRUE)
}
\arguments{
\item{spe}{the SpatialExperiment from EH7312}

\item{phv}{character() vector of phenotype labels, from colData}

\item{add_cd4}{logical(1) if TRUE, use CD3+ CD8- to enumerate CD4}

\item{add_multicol}{logical(1) if TRUE, add a multilevel factor for cell type}
}
\description{
Explore ovVP Vectra Polaris data from ExperimentHub EH7312
}
\note{
App permits selection of specific tissue slices on tissue microarray, coloring
the points at cell coordinates in `spatialCoords`.  The input SpatialExperiment must
include a column 'nid' of numerical slice ids in colData.
}
\examples{
if (interactive()) {
  requireNamespace("ExperimentHub")
  ovVP = ExperimentHub::ExperimentHub()[["EH7312"]]
  ovVP$nid = as.numeric(factor(ovVP$sample_id))
  display_by_core(ovVP, ovVP_phv())
}
}
