% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{litov}
\alias{litov}
\title{small subset of a single core from EH7312}
\format{
SpatialExperiment
}
\usage{
litov
}
\description{
small subset of a single core from EH7312
}
\note{
Retains only 3000 cells.
}
\keyword{datasets}
