% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{luV3_phv}
\alias{luV3_phv}
\title{return a vector of colData labels denoting cell phenotypes for EH7311}
\format{
An object of class \code{character} of length 6.
}
\usage{
luV3_phv
}
\description{
return a vector of colData labels denoting cell phenotypes for EH7311
}
\keyword{datasets}
