% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ovVP_phv}
\alias{ovVP_phv}
\title{return a vector of colData labels denoting cell phenotypes for EH7312}
\usage{
ovVP_phv()
}
\description{
return a vector of colData labels denoting cell phenotypes for EH7312
}
