% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovVP_summaries}
\alias{ovVP_summaries}
\title{table of sums of phenotype counts collapsed over cells within tissue microarray slices}
\format{
data.frame
}
\usage{
ovVP_summaries
}
\description{
table of sums of phenotype counts collapsed over cells within tissue microarray slices
}
\note{
Produced using get_summaries
}
\keyword{datasets}
