% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{show_core}
\alias{show_core}
\title{use ggplot to display cell phenotype over positions}
\usage{
show_core(spe, colvar = "phenotype_cd8", ...)
}
\arguments{
\item{spe}{SpatialExperiment instance}

\item{colvar}{character(1) phenotype label}

\item{\dots}{passed to geom_point}
}
\description{
use ggplot to display cell phenotype over positions
}
\examples{
data(litov)
show_core(litov)
}
